+++ packages/substrate-bindings/src/index.ts
@@ -2,3 +2,4 @@ export * from "./codecs"
 export * from "./hashes"
 export * from "./storage"
 export * from "./types"
+export * from "./utils"
